Add example project files
Followed the [creating a new project][1] steps in the Next.js tutorial.
Amended the command to create the project to use pnpm and eslint, and
to make it explicit that we are using typescript.

Then had to move all the project files that were created up one
directory into the root directory (could not create the project directly
in the root directory as it said there was a conflict with the existing
files).

[1]: https://nextjs.org/learn/dashboard-app/getting-started#creating-a-new-project